refactor: change PayPal Checkout to PayPal Donations
This also changes the PayPalCheckout namespace to PayPalCommerce to keep it clear as to which API it reflects
